refactor(ucp): split adapter impl into bounded modules (wave8b step2)
